watcom compilation fix; for some reason this failed.

+++ src/games/dekware/psicon/wincon.c
@@ -1226,7 +1226,7 @@ int CPROC SetFlash( PDATAPATH pdp, PSENTIENT ps, PTEXT parameters )
 // methods for window logic routines to use as callbacks...
 //----------------------------------------------------------------------------
 
-static void CPROC DrawString( PCONSOLE_INFO pmdp, int x, int y, RECT *r, TEXTCHAR *s, int nShown, int nShow )
+static void CPROC DrawString( PCONSOLE_INFO pmdp, int x, int y, RECT *r, TEXTCHAR *s, size_t nShown, size_t nShow )
 {
    SIZE size;
 	//lprintf( WIDE("Adding string out : %s %d %d at %d,%d or (%d,%d)  (%d,%d)"), s, nShown, nShow,x,y,r->left,r->top, r->right, r->bottom );
